refactor: extract shared useFetchPage hook to deduplicate 4 skins
Extract ~420 lines of duplicated fetch logic from 4 FetchPage.tsx skins
into a shared useFetchPage() hook in panel/src/core/hooks/useFetchPage.ts.

Exports: useFetchPage, Provider, FetchState, MAX_URLS, isSafeUrl, parseUrls
Net reduction: -236 LOC with single source of truth for all fetch behavior.
