+++ panel/src/skins/carbon/pages/FetchPage.tsx
@@ -2,28 +2,19 @@
  * 网页抓取页面 - Carbon 皮肤版本
  *
  * 文件用途：提供网页内容抓取界面，支持多 URL 输入、多种抓取提供商、结果展示和导出
+ *
+ * 业务逻辑统一抽取至 `@core/hooks/useFetchPage`，本文件仅保留 Carbon 皮肤特有的 UI 渲染。
  */
 
-import { useState, useCallback, useEffect, useRef, type FormEvent } from 'react'
-import { useTranslation } from 'react-i18next'
 import { m } from 'framer-motion'
 import {
   Globe, ExternalLink, Copy, Download, CheckCircle2,
   AlertCircle, ChevronDown, ChevronUp, Terminal,
   Settings, Link as LinkIcon,
 } from 'lucide-react'
-import { api } from '@core/services/api'
-import { useNotificationStore } from '@core/stores/notificationStore'
-import { formatError } from '@core/lib/errors'
-import type { FetchResponse, FetchResult } from '@core/types'
+import { useFetchPage, isSafeUrl, type Provider } from '@core/hooks/useFetchPage'
 import { staggerContainer, staggerItem, fadeInUp } from '@core/lib/animations'
 import styles from './FetchPage.module.scss'
-
-const isSafeUrl = (url: string): boolean => /^https?:\/\//i.test(url)
-
-const MAX_URLS = 20
-
-type Provider = 'builtin' | 'jina_reader' | 'tavily' | 'firecrawl' | 'exa'
 
 const PROVIDERS: { value: Provider; label: string; description: string }[] = [
   { value: 'builtin', label: 'BUILTIN', description: 'Built-in fetcher' },
@@ -33,135 +24,28 @@ const PROVIDERS: { value: Provider; label: string; description: string }[] = [
   { value: 'exa', label: 'EXA', description: 'Exa.ai API' },
 ]
 
-type FetchState =
-  | { status: 'idle'; message: null }
-  | { status: 'loading'; message: null }
-  | { status: 'error'; message: string }
-
 export function FetchPage() {
-  const { t } = useTranslation()
-  const [urls, setUrls] = useState('')
-  const [provider, setProvider] = useState<Provider>('builtin')
-  const [timeout, setTimeout_] = useState(30)
-  const [showAdvanced, setShowAdvanced] = useState(false)
-  const [fetchState, setFetchState] = useState<FetchState>({ status: 'idle', message: null })
-  const [results, setResults] = useState<FetchResponse | null>(null)
-  const [expandedItems, setExpandedItems] = useState<Set<number>>(new Set())
-  const addToast = useNotificationStore((s) => s.addToast)
-  const activeRequestRef = useRef<AbortController | null>(null)
-  const inputRef = useRef<HTMLTextAreaElement>(null)
-
-  useEffect(() => {
-    return () => {
-      activeRequestRef.current?.abort()
-    }
-  }, [])
-
-  const parseUrls = (text: string): string[] => {
-    return text
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line.length > 0 && (line.startsWith('http://') || line.startsWith('https://')))
-  }
-
-  const canFetch = parseUrls(urls).length > 0
-  const isLoading = fetchState.status === 'loading'
-
-  const handleFetch = useCallback(
-    async (e: FormEvent) => {
-      e.preventDefault()
-      if (!canFetch) return
-
-      const urlList = parseUrls(urls)
-      if (urlList.length === 0) {
-        addToast('error', t('fetch.noValidUrls', 'No valid URLs found'))
-        return
-      }
-      if (urlList.length > MAX_URLS) {
-        addToast('error', t('fetch.tooManyUrls', { max: MAX_URLS, count: urlList.length }))
-        return
-      }
-
-      activeRequestRef.current?.abort()
-      const controller = new AbortController()
-      activeRequestRef.current = controller
-
-      setFetchState({ status: 'loading', message: null })
-      setResults(null)
-      setExpandedItems(new Set())
-
-      try {
-        const res = await api.fetch(urlList, provider, timeout, controller.signal)
-        setResults(res)
-        setFetchState({ status: 'idle', message: null })
-        addToast('success', t('fetch.success', { count: res.total_ok, total: res.total }))
-      } catch (err) {
-        if (controller.signal.aborted) return
-        const message = formatError(err)
-        setFetchState({ status: 'error', message })
-        addToast('error', t('fetch.failed', { message }))
-      } finally {
-        if (activeRequestRef.current === controller) {
-          activeRequestRef.current = null
-        }
-      }
-    },
-    [urls, provider, timeout, canFetch, addToast, t],
-  )
-
-  const handleRetry = useCallback(() => {
-    const syntheticEvent = { preventDefault: () => {} } as FormEvent
-    handleFetch(syntheticEvent)
-  }, [handleFetch])
-
-  const toggleExpanded = (index: number) => {
-    setExpandedItems((prev) => {
-      const next = new Set(prev)
-      if (next.has(index)) {
-        next.delete(index)
-      } else {
-        next.add(index)
-      }
-      return next
-    })
-  }
-
-  const copyToClipboard = async (text: string) => {
-    try {
-      await navigator.clipboard.writeText(text)
-      addToast('success', t('fetch.copied', 'Copied to clipboard'))
-    } catch (err) {
-      addToast('error', t('fetch.copyFailed', 'Failed to copy'))
-    }
-  }
-
-  const downloadAsMarkdown = (item: FetchResult) => {
-    const md = `# ${item.title || item.url}\n\n**URL:** ${item.final_url}\n\n${item.content || ''}`
-    const blob = new Blob([md], { type: 'text/markdown' })
-    const url = URL.createObjectURL(blob)
-    const a = document.createElement('a')
-    a.href = url
-    a.download = `${item.title?.replace(/[^a-zA-Z0-9]/g, '_') || 'content'}.md`
-    a.click()
-    URL.revokeObjectURL(url)
-    addToast('success', t('fetch.downloaded', 'Downloaded as Markdown'))
-  }
-
-  const exportAllAsMarkdown = () => {
-    if (!results || results.results.length === 0) return
-    const md = results.results
-      .filter((item) => !item.error)
-      .map((item) => `# ${item.title || item.url}\n\n**URL:** ${item.final_url}\n\n${item.content || ''}\n\n---\n`)
-      .join('\n')
-    const blob = new Blob([md], { type: 'text/markdown' })
-    const url = URL.createObjectURL(blob)
-    const a = document.createElement('a')
-    a.href = url
-    a.download = 'all_content.md'
-    a.click()
-    URL.revokeObjectURL(url)
-    addToast('success', t('fetch.allDownloaded', 'All content downloaded'))
-  }
+  const {
+    t,
+    urls, setUrls,
+    provider, setProvider,
+    timeout, setTimeout_,
+    showAdvanced, setShowAdvanced,
+    fetchState,
+    results,
+    expandedItems,
+    inputRef,
+    validUrls,
+    canFetch,
+    isLoading,
+    hasResults,
+    handleFetch,
+    handleRetry,
+    toggleExpanded,
+    copyToClipboard,
+    downloadAsMarkdown,
+    exportAllAsMarkdown,
+  } = useFetchPage()
 
   const renderResults = () => {
     if (isLoading) {
@@ -326,8 +210,6 @@ export function FetchPage() {
     )
   }
 
-  const hasResults = results !== null
-
   return (
     <div className={styles.page}>
       <div className={styles.gridOverlay} />
@@ -364,7 +246,7 @@ export function FetchPage() {
               required
             />
             <div className={styles.urlCount}>
-              {parseUrls(urls).length} {t('fetch.validUrls', 'valid URLs')}
+              {validUrls.length} {t('fetch.validUrls', 'valid URLs')}
             </div>
           </div>
